LayoutTests/http/tests/security/cross-frame-access-protocol-explicit-domain-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL https://127.0.0.1:8443/security/resources/cross-frame-iframe-with-explicit-domain-set.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-protocol-explicit-domain.html. Domains, protocols and ports must match.

This test currently fails because we check the port and protocol even if document.domain is explicitly set (rdar://problem/5366437).


PASS: Cross frame access to https from http, after explicitly setting document.domain, was denied.


--------
Frame: 'aFrame'
--------
Inner iframe with explicit document.domain set.
